Do the URL handling and search redirect locally.
Instead of having to send an actual search request, just let the
webextension figure out if the search terms are a URL or not. If they
are, redirect to the search term as a normalized URL. If not, use the
configured search provider URL, which defaults to DuckDuckGo. Bump the
version to 1.2.